visualization-app/requirements.txt
streamlit==1.29.0
networkx==3.2.1
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
google-generativeai==0.3.1
python-dotenv==1.0.0